--- packages/runner-ct/static/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-      <!-- Uncomment to connect using standalone react-devtools -->
-      <!-- <script src="http://localhost:8097"></script> -->
-      <meta http-equiv="content-type" content="text/html;charset=utf-8" />
-      <meta name="viewport" content="width=device-width, initial-scale=1.0">
-      <title>{{projectName}}</title>
-
-      <link href="/{{namespace}}/runner/favicon.ico?v2" rel="icon">
-
-      <link rel="stylesheet" href="/{{namespace}}/runner/cypress_runner.css">
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="text/javascript" src="/{{namespace}}/runner/cypress_runner.js"></script>
-    <script type="text/javascript">
-      // set a global so we know the 'top' window
-      window.__Cypress__ = true
-    </script>
-  </body>
-</html>
