--- packages/runner-ct/test/.mocharc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  spec: 'src/**/*.spec.ts',
-}
