--- packages/runner-ct/test/setup.js
@@ -1,1 +1,0 @@
-global.window = {}
