+++ packages/runner-shared/src/index.ts
@@ -1,5 +1,3 @@
 export * from './dom'
 
 export * from './studio'
-
-export * from './store'
