--- packages/runner-shared/src/mixins.scss
@@ -1,4 +1,0 @@
-@mixin button-active {
-  background-color: #e9e9e9;
-  box-shadow: inset 0 3px 5px rgba(0, 0, 0, 0.125);
-}
