--- packages/runner-shared/src/selector-playground/selector-playground.scss.d.ts
@@ -1,7 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-
-}
-export const cssExports: CssExports;
-export default cssExports;
