--- packages/runner-shared/src/styles.module.scss
@@ -1,10 +1,0 @@
-.specsList {
-  position: relative;
-  display: grid;
-  grid-template-rows: auto 1fr;
-  height: 100%;
-  min-width: 150px;
-
-  background: white;
-  z-index: 0;
-}
