--- packages/runner-shared/src/variables.scss
@@ -1,21 +1,0 @@
-$message-height: 33px;
-
-%clearfix {
-  &:before,
-  &:after {
-    content: "";
-    display: table;
-  }
-  &:after {
-    clear: both;
-  }
-}
-
-$success: #08c18d;
-$error: #e94f5f;
-$link-text: #3380FF;
-$reporter-min-width: 450px;
-$message-height: 33px;
-$font-mono: Consolas, Monaco, 'Andale Mono', monospace;
-$font-sans: 'Mulish', 'Helvetica Neue', 'Arial', sans-serif;
-$open-sans: 'Open Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
