+++ packages/runner-shared/test/utils.ts
@@ -2,7 +2,6 @@ import $Cypress from '@packages/driver'
 import { EventManager } from '@packages/app/src/runner/event-manager'
 import type { Socket } from '@packages/socket/lib/browser'
 import { StudioRecorder } from '../src/studio'
-import { selectorPlaygroundModel } from '../src/selector-playground'
 import * as MobX from 'mobx'
 
 export const StubWebsocket = new Proxy<Socket>(Object.create(null), {
@@ -15,7 +14,7 @@ export const createEventManager = () => {
   return new EventManager(
     $Cypress,
     MobX,
-    selectorPlaygroundModel,
+    {}, // TODO: Bring back "Cypress Studio" and integrate with 10.x runner // selectorPlaygroundModel
     StudioRecorder,
     StubWebsocket,
   )
